Validate during Build creation the following:
- If the Strategy exist, either ClusterBuildStrategy or
  BuildStrategy
- If the Build Output Credentials are set, validate if the
  secret exist on the Build namespace

For making a proper unit-test validation of the above, this
PR includes the following:
- Ginkgo suite to test the controller build package
- A build_controller_test.go with all related scenarios for the
  above mentioned validations
- Fakes for the controller-runtime manager and client, hosted under
  pkg/controller/fakes. This allow us to use this fakes to test actions
  of the Build controller during reconciliation.
- Added a new catalog.go class under the test folder, this allow us to host
  functions that we can use during the unit tests, to keep the code clean and
  avoid repetitions.

The fake classes are generated with counterfeiter(see https://github.com/maxbrunsfeld/counterfeiter)
I added a make target so that developers can trigger this when neededm and a related script.

I modified the "make test" target to use ginkgo and to avoid using the vendor directory,
this should be already covered when setting the GO111MODULE to on.