Add env variables for the e2e tests (#87)
* Update HACK.md

Add instructions on how to run e2e tests with
environment variables for a customize registry

* Introduce env vars when running e2e tests

This enable the user to keep using the
Build CRS instances examples, but with a
customize registry. Not the default openshift one.

This allows the user to provide three env variables:
- A registry namespace
- A registry endpoint
- A registry secret

* Simplified e2e tests

Adding a new set of test cases, specially with a different folder
for the Build example turn out to be not that easy, while this
implied a lot of code repetition.

This commit simplifies the approach so that future new e2e test
cases can easily be integrated inside the main_test.go

- Adding a helper struct to store different objects and data require
  for running assertions on the e2e
- Remove code that was repeated in different test classes and that
  could be all compacted inside the validators.go

* Add e2e tests for private github/gitlab repositories

- update HACK.md readme
- Add test/date directory, to host examples for private repositories
- Add new set of ENV variables required to customize the e2e tests
- Avoid running private repos e2e test if a flag is not set(_for not breaking CI_)

* Simplify env variables for private git repos

Only rely on 3 vars,
- TEST_WITH_PRIVATE_REPO for running the e2e tests with private repos
- TEST_PRIVATE_GITHUB for defining the github private endpoint
- TEST_PRIVATE_GITLAB for defining the gitlab private endpoint

Also, added gitlab testcase for Kaniko

* Ensure buildpacks v3 tests are run only if the env variables
for the registry are set. If not, skip them.