Add an Status for the Build controller
The Status comes in the form of `Registered` , exposing
the status of the controller after Reconciliation. This uses
the corev1.ConditionStatus, therefore only three values can take place:
- true
- false
- unknown

If the Build instance fails to reconcile due to missing resources defined
in the instance, the Registered status will became `false`, if the validations
are fine, then the status will become `true`.

Also adding a fake class for the client StatusWriter interface, this allow us
to stub the controller status depending on the test case. This come with new
assertions on the unit test and a new func on the catalog.go helper class.