Validate during Build creation the following: (#99)
* Validate during Build creation the following:

- If the Strategy exist, either ClusterBuildStrategy or
  BuildStrategy
- If the Build Output Credentials are set, validate if the
  secret exist on the Build namespace

For making a proper unit-test validation of the above, this
PR includes the following:
- Ginkgo suite to test the controller build package
- A build_controller_test.go with all related scenarios for the
  above mentioned validations
- Fakes for the controller-runtime manager and client, hosted under
  pkg/controller/fakes. This allow us to use this fakes to test actions
  of the Build controller during reconciliation.
- Added a new catalog.go class under the test folder, this allow us to host
  functions that we can use during the unit tests, to keep the code clean and
  avoid repetitions.

The fake classes are generated with counterfeiter(see https://github.com/maxbrunsfeld/counterfeiter)
I added a make target so that developers can trigger this when neededm and a related script.

I modified the "make test" target to use ginkgo and to avoid using the vendor directory,
this should be already covered when setting the GO111MODULE to on.

* Add documentations for unit-test

- Insights on ginkgo and counterfeiter

Improve a function code, and move two lines
into a one-liner

* Add an Status for the Build controller

The Status comes in the form of `Registered` , exposing
the status of the controller after Reconciliation. This uses
the corev1.ConditionStatus, therefore only three values can take place:
- true
- false
- unknown

If the Build instance fails to reconcile due to missing resources defined
in the instance, the Registered status will became `false`, if the validations
are fine, then the status will become `true`.

Also adding a fake class for the client StatusWriter interface, this allow us
to stub the controller status depending on the test case. This come with new
assertions on the unit test and a new func on the catalog.go helper class.