Add proposal for the BuildStrategy API
Even though we already have the BuildStrategy API in use, this proposal is a brief vision document of the initial goals of the BuildStrategyAPI.

At the timing of writing this proposal, parts of it has already been implemented and is in master. Ideally, a proposal should be as focused as possible. This proposal makes an exception to that since I'm proposing the proposal process after significant functionality has already been built.

Co-authored by @qu1queee