Introduce config package
This should help in the future to collocated different
configuration the Build controllers might need.

Currently this only contains a single field, CtxTimeOut.

The CtxTimeOut is a field that allow us to define the time out in
seconds we want to have for all controllers Kubernetes API calls.

If a Kubernetes API call exceed this timeout, the call will be aborted.
This helps to avoid overloading the Kubernetes API with calls that might
be hanging.

The CtxTimeOut have a default value of 300 seconds.