Improve logging and add proper go contexts (#248)
* Add new ctxlog package

This will allow us to properly get a logging instance
from the main.go to all the different set of controllers.

Also, this package allow us to properly use a context, so that
when we do k8s API calls, we can define a proper context with
a timeout, and not a context.TODO

* Use the ctxlog package

For initializing the logger inside the `main.go`.
In each controller class.

* For the ctxlog pkg

When extracting a logger, if there is no matching value
with the key, return a no-op logger, that does nothing

* Fix Error logging in Build Controller

When logging we were not sending a key pair,
but only a key.

* Ensure unit-test are passing

Unit-tests for controllers required to
use a no-op logger.

* Introduce config package

This should help in the future to collocated different
configuration the Build controllers might need.

Currently this only contains a single field, CtxTimeOut.

The CtxTimeOut is a field that allow us to define the time out in
seconds we want to have for all controllers Kubernetes API calls.

If a Kubernetes API call exceed this timeout, the call will be aborted.
This helps to avoid overloading the Kubernetes API with calls that might
be hanging.

The CtxTimeOut have a default value of 300 seconds.

* Make use of the new config package

Ensure during the controllers initialization that a proper config
is instantiated. Also, ensure that during each controller reconciliation
we are using the context timeout from the Config instance.