Fix readiness/liveness probe when HA is enable (#276)
When running in HA mode, the previous localhost:8383/metrics
endpoint was not available for pods that are not leaders. This
end-up in pods having a CrashLoopBackOff  STATUS.

In order to fix this, we are introducing a different approach on
how we check the Readiness/Liveness by using the operator-sdk ready
package.

Also, we test this manually, it should be working when HA is enabled.

Signed-off-by: EEE <encalada@de.ibm.com>