BuildRun reconciles enhancement (#313)
* Improve BuildRun Reconciler logic

This should provide improvements on the controller performance.

The main problem is that we are reconciling for every change on a
TaskRun StartedTime even if the REASON is not changed. This is leading
too a large amount of reconciliations in the BuildRun(> 10) for a single
Build, plus all the API calls we do per Reconcile.

This PR improves this by using a more fine grained set of Predicates when
watching TaskRuns and also it make use of the EnqueueRequestsFromMapFunc that
generates the List of Reconciles we need to do per event, instead of Reconciling
for every event.

* Comment out some print statements

These are polluting the output of the test-unit make target.

* Check for errors when a taskrun does not exist
Move the closing reconcile logging statement into
the right place

* Extend test-cases for the BuildRun controller

This adds missing unit-tests to ensure that for every TaskRun conditions
update, our BuildRun gets the right values under its Status fields

* Address PR reviews

Do not list during the filtering of events for Reconciles, but
rather just return an slice with a TaskRun if the proper labels are
in place.

Simplify repeated code into a single variable.

Fix some functions with the wrong names