Rollback to use a generated name for each TaskRun
Similar to how a Deployment resource generates names for pods, we
should stick to an auto-generated name for TaskRuns. This will
avoid issues like #315.

We use to have the auto-generated name before, but it was modified
with PR #313

With different names between BuildRuns and TaskRun, some code refactoring
on the buildrun controller was required to preserve the improvements fone
in PR #313. Also, unit-tests were improved to deal with different Reconcile
request on the BuildRun controller, depending if they come from a BuildRun
object or a TaskRun object event.