For BuildRun update events
Do not reconcile if a BuildRun is already referencing a TaskRun. Reconciling
on Update events of the BuildRun will end-up on generating new TaskRuns.

The above also should cover the case when a TaskRun is created but is not yet
available due to some high load on k8s, while another reconcile request takes place
and tries to generate a new TaskRun.