Rollback to use a generated name for each TaskRun (#323)
* Rollback to use a generated name for each TaskRun

Similar to how a Deployment resource generates names for pods, we
should stick to an auto-generated name for TaskRuns. This will
avoid issues like #315.

We use to have the auto-generated name before, but it was modified
with PR #313

With different names between BuildRuns and TaskRun, some code refactoring
on the buildrun controller was required to preserve the improvements fone
in PR #313. Also, unit-tests were improved to deal with different Reconcile
request on the BuildRun controller, depending if they come from a BuildRun
object or a TaskRun object event.

* Ensure BuildRun is properly updated when a TaskRun
is deleted but not the BuildRun.

Addressed some of the PRs comments

* For BuildRun update events

Do not reconcile if a BuildRun is already referencing a TaskRun. Reconciling
on Update events of the BuildRun will end-up on generating new TaskRuns.

The above also should cover the case when a TaskRun is created but is not yet
available due to some high load on k8s, while another reconcile request takes place
and tries to generate a new TaskRun.

* Improve regexp function

generate a regexp variable and reuse that
for any other operation