Initial Release Pipeline
Creating an initial image release pipeline for the shipwright operator.
This makes use of the travis docker service to build the image and push
it to quay.io when release tags are made and PRs merge in the main
branch.

In the future we will want to dogfood our own operator and use
Shipwright to build and push the image.