Add ramp-up duration metrics to build (#365)
When building, a couple of Kubernetes resources are created to run the
respective build strategy to compile the app. Depending on the system
load or utilisation, it may take more or less time to create the
resources to kick the run. The time to ramp-up a specific part of the
build can be useful for monitoring and alerting.

Add three ramp-up duration metrics to `build`:
- between buildrun creation and taskrun creation (buildrun ramp-up)
- between taskrun creation and taskrun pod creation (taskrun ramp-up)
- between pod creation and last init container completion (pod ramp-up)

Note: The ramp-up durations usually range between zero (no measurable)
and one second. Therefore, the default bucket range is simply counting
up from zero to ten seconds.