Refactor metrics package unit tests
The PR #365 contained a copy and paste error that slipped through the
fact that unit tests in that package exist. The unit tests did check
for specific metric variables by directly addressing them, because the
unit tests were in the same package as the metrics code. This is also
something that goes against typical test setups.

Move test code into a test package.

Refactor test code to not use the direct metric variables, but by
requesting the metrics via the registry.

Add setup code to simulate a couple of buildruns, which generate metrics
that can be checked in the `It` clauses.

Fix copy and paste error where the wrong metrics were used.

Add all metrics that are defined in the metrics package to the registry.