Enhance docs by adding contributing and code-of-conduct files (#372)
* Add code of conduct doc

We reference to the Kubernetes community one, in the same way
as most of the Kubernetes existing projects

* Add contributing doc

Trying to add some initial words on future contributions. I think we
still need to generate more docs for this, but so far is good to have
this as an initial step.

* add copyright headers