Remove ruby samples for Buildpacks
Due to our default CBS being Paketo, we should not have samples
for ruby buildpacks. Paketo is not officially supporting this
runtime yet, but there is a proposal already https://github.com/paketo-buildpacks/rfcs/blob/master/accepted/0005-ruby-promotion.md

I would like to remove this from the repo for now, because it can lead to confusion. We can build
the image for ruby, but during deployment times, this will not work due to missing dependencies. This
could lead to bad experiences for the Build.

I would bring this samples back once the gcr.io/paketo-buildpacks/builder:latest have support
for Ruby buildpacks.