Introduce namespace build-operator lock configmap
The current namespace of the build-operator lock configmap is hard-coded
to be `default` at the moment. With the new override possibilities, it
can be set to a more suitable location.

Use downward API to set `BUILD_OPERATOR_LEADER_ELECTION_NAMESPACE`
environment variable in the deployment YAML for the build-operator to be
the same namespace as the operator pod is located in.