Fix go module files
The go.mod does not have a proper indentation for the require() block.
The go.mod was missing the changes introduced for the k8s.io/utils, this
were already added in the vendor and the vendor/modules.txt before.

The go.sum was not reflecting all changes.

I detected the above by a simple `go mod tidy`.