Report build metrics as soon as possible
There are metrics that are observed when the build run is finished even
though the actual details were available already. Therefore, it could
happen that some metrics are *not* reported due to a failing build run
and would be lost. If the metrics observation would be right after the
creation of the `TaskRun`, then the metrics should be reported in any
case and should also only be reported once.

Move `BuildRunCountInc` and `BuildRunRampUpDurationObserve` calls to
after when the `TaskRun` is created, because at that point in time the
required information are already available.

Move `BuildRunEstablishObserve` to the section of the code, where the
required time fields are set for the first time.