Conditions framework implementation
- Adding a generic conditions pkg, so that any controller that needs
to support Conditions can use the pkg to set/retrieve their related
conditions.

- Adding a Condition of the type Succeeded for resources that run to
completion, like the BuildRun

- Adding a generic Condition types. This can be extended in the future
with new fields.

- Adding unit tests for the new conditions pkg.

- Adding support for the Conditions pkg in the BuildRun, although the
CRD is not yet generating or updating the Status.Condition object. This
is done by extending the Status struct to include a Conditions field.

The conditions framework is heavily based on
https://github.com/knative/pkg/blob/master/apis/condition_set.go