tests: Get BuildRun before parent Build is deleted
When testing if a BuildRun is cleaned up by k8s garbage collection, we
need to get the BuildRun object before we delete the parent Build
object. Our test cluster in Travis probably didn't catch this because of
weak resource constraints. This test fails when run against a
production-grade cluster.