Fix buildah ClusterBuildStrategy
For quay.io/buildah/stable:latest, the push cmd changed the arguments
behaviour. It now requires an IMAGE and a DESTINATION. The DESTINATION
can be ommited, and buildah would add that based on the IMAGE, by adding
a docker transport prefix (`docker://`).

Add IMAGE name to command explicitely and keep DESTINATION as-is.

Signed-off-by: Matthias Diester <matthias.diester@de.ibm.com>