Fix flaky unit-test spec for Build
We assert for a combination of secret names under Status.Message when
the secrets validation fails. Because the list of secrets comes from a
map and a map is an unordered collection of key-value pairs, this
commits adds a sort, to ensure an ordered list in Status.Message and for
avoiding the flakiness in the unit-test spec.