Tweak build operator build settings
There are linker flags that can make our build operator binary and
therefore our container image smaller. For example, we can omit the
symbol table and debug information and the DWARF symbol table. In case
of the build operator this means we can save 20% of the regular size.

With `ko` having no way of giving it `-ldflags` directly, we have to
rely on the GOFLAGS environment variable. However, this comes with the
problem that it does not support flags with values that contain spaces,
because they use space as a separator. This means we can only pick one
flag value for `-ldflags`.

Add `-ldflags=-s` equivalent to the Go build flags.