Remove build dependency in local Make targets
For Makefile target to run the operator locally, there is no need to build
the operator first. In fact, the operator SDK local mode builds the binary
itself as part of the local run mode. Removing the `build` speeds up the
start time and avoids confusion towards which binary is used in local mode.

Remove `build` dependency and replace it with a simple `vendor` dependency.