Fix nil pointer dereference in taskrun processing
There seems to be a rather rare case where we have a setup in which a TaskRun
has a completion timestamp and condition reason `TaskRunReasonFailed`, while
the respective pod has containers which are not terminated, (yet). In this
case the code ends up with a nil pointer dereferences and crashes the running
operator pod. See #591 for details.

Change `updateBuildRunUsingTaskRunCondition` function to set the failed
container name only if a failed container could be looked-up.

Add unit-test to construct a represenation of the error scenario and test that
the `Reconcile` runs through without a nil pointer dereference.