This replaces the taxi go app endpoint
The existing one have an issue on the go.mod definition, leading
to build failures when using the latest golang image.

This PR changes the repo to use, which have the fixes in place. This
is a fix PR, the end-goal is to tackle issue #366 which should avoid
problems like this in the future.