Remove scaffold code from operator-sdk init
This code was generated as part of the scaffold when using the sdk. This
code was intended to mitigate a user error observed when using
"operator-sdk up local" cmd.

See https://github.com/operator-framework/operator-sdk/pull/2190 for
more information.

We do not longer need this code. The metrics endpoint for prometheus is
initialized by controller-runtime in the background.

Signed-off-by: Matthias Diester <matthias.diester@de.ibm.com>