Remove operator SDK package pkg/test
The current end to end test heavily rely on the `pkg/test` package by
using the so called `framework` package. In order to remove the operator
SDK package dependency and make test end to end test closer to the
integration test, switch the implementation to use as many integration
test utility code as possible by also keeping the test cases as close to
the current one as possible.