Fixes #518
The behaviour for a git repo revision was being hardcoded to `master`,
which is incorrect, while github is already creating the default branch
to `main` and also because the default Tekton behaviour does not take
place if the revision parameter is not specified.

Tekton defaults behaviour triggers a git symbolic-link to HEAD, which
we should allow to happen if the user does not specify a revision in the
Build source.