Refactor preparations prior to taskRun generation
Simplify the createTaskRun func()
Simplify the strategies retrieval and make it a standalone step prior
to the TaskRun spec generation.
Ensure the serviceAccount retrieval is a standalone step prior to the
TaskRun spec generation.
For both serviceAccount and strategies retrieval, fail the TaskRun if
any of the custom checks take place or if a referenced object is not
found, otherwise allow one more reconciliation on sys call failures.

Signed-off-by: Matthias Diester <matthias.diester@de.ibm.com>