Enhance failures on buildRun
On Create predicate events, do not reconcile if the CompletionTime is
set for BuildRuns.

When retrieving a Build object, reconcile again if is a sys call or
mark the BuildRun as failed if the object was not found.

When the Build object registration failed, mark the BuildRun as failed.

When creating the TaskRun, allow to reconcile one more time on errors.

Remove obsolete functions for updating BuildRun conditions.

Add unit-tests for Build object retrieval.