Fix nil-pointer dereference of BuildSpec Strategy
There were cases in which the Build had a `nil` Strategy, which is copied into
the BuildRun `Status.BuildSpec` and later used internally, for example the
metrics code is using the `Strategy` field. In case it is `nil`, the function
call for the metrics failed with a nil-pointer dereference.

Introduce `BuildSpec` convenience function called `StrategyName` to return the
name of the strategy, which includes a `nil` safety check.

Replace all occurrences of `BuildSpec.Strategy.Name` with the newly introduced
function to avoid panics.