Ensure GO_OS does not us an OS that isn´t "linux"
At the moment the only supported OS for building the image of the
controller is "linux". When running targets like "make install-controller",
in a macOS, GO_OS is set to "darwin" which fails building the image.

This PR ensures that if GO_OS is not set, we default to "linux". This
flexibility is not needed, but might be needed in the future. I added
the usage of the origin function on purpose, to make the logic very clear.