Use imjasonh/setup-ko GitHub Action
See https://github.com/imjasonh/setup-ko

This action installs the latest `ko` release, so we don't need to bump
the version ourselves every time there's a release.

We can also pin to a ko release if we want to, to avoid a breakage or
bad version.

Referencing the action by SHA (equivalent to v0.3) to avoid silent
behavior changes and breakages.