Skip Git LFS tests in case there is no Git LFS
In the first version of `cmd/git`, it was assumed that Git LFS is installed and
used in the unit tests. As not all environments might have Git LFS, it would
lead to unnecessary unit test failures.

Refactor entry code to only run environment check when executed as the tool.

Add skip code to unit tests to omit tests in case Git LFS is not installed.