Add settings to disable user prompts
There seem to be cases where depending on the system and configuration
the Git CLI prompts for username and password. We do not want the
wrapper tool to ever prompt the user as this is breaking the process.

Add `BatchMode` to SSH options to disable prompts.

Set `GIT_TERMINAL_PROMPT` environment variable to disable prompts.

Connect `/dev/null` to Git command (via `nil`) to disable prompts.