Mitigate context deadline exceeded issue
With the default implementation of `repo.List`, the timeout seems to be rather
short and there are intermittent context deadline exceeded errors when running
the unit tests under load, or when repeating them over and over again.

Mitigate the error by switching to the list function that accepts a context. It
was added recently and privides the possibility to define a different timeout.

Refactor related code to make it more compact by removing superfluous returns.