Fix Git submodule update issue with SSH creds
The credential look-up routine only set up the Git command line flags for the
Git `clone` command, but not for the subsequent `submodule` command. When a
user tried to use a repository with a submodule that required authentication as
well, the Git step failed.

Change code to store the respective credential command line arguments into a
separate slice so that it can be used for different Git commands.