Check for Tekton installed before running the manager
Right now reconciler tries to watch TaskRun objects when manager
starts. If tekton is not intalled, controller crashloops.
This change brings check for TaskRuns existing in the earlier
stage and exits with explicit error and explanation.
It also writes to the termination log file to propagate the
error to the Status of the pod.
This way user should know perfectly well what the problem is
and how to fix it.