Improve Documentation for build and buildRun
We should make sure that examples in the documentation for the build CDR have
correct examples regarding the properties that the want to illustrate and that
links to samples are kept up-to-date.

Modified example in `build.md` that should illustrate the revision property and
fixed relative paths to build and buildRun config for `buildstrategies.md`.