Mitigate bundle E2E test build create issue
Depending on the test infrastructure, it can take a bit of time until the
build resource is available in the cluster. If a buildrun starts too fast,
it can happend that it does not find the build yet.

Add look-up in `Create` to verify that the create operation is completed.