Add check to verify `--depth` can be used
For older Git versions, the `git submodule update` command does not have
the `--depth` flag to create shallow clones, instead the usage of `--depth`
leads to errors trying to update submodules.

Add check to verify that `--single-branch` is available in `update` as this is
a clear indication that shallow clones for submodules are available.

Group these help inspections in a separate function.