Update Verify CI job to check CRD generation
We have a Verify CI job that runs the basic static analysis, like code linting. This step should also ensure that CRD definitions for the Build APIs are aligned with what is in the golang code.
Fixes #943

Update hack/verify-generate.sh to check CRDs and deep copy generation
Update deep copy generation file not in sync