Move Git URL rewrite into private key section
The Git URL rewrite logic was located outside of the credentials section of the
code, which means it was used in any case. However, it does not make sense to
be used in combination with a HTTPS URL and basic auth.

Move section as-is into the private key section and setup.