Fix accessing non-terminated steps for failures
The cause of this bug is the assumption that `v1beta1.TaskRunReasonFailed` means that
all steps are terminated. However, it only guarantees that at least on
step is terminated with non-zero exit code.

Adds extra non-terminated step to test to capture the scenario above.

Adds new condition to avoid accessing non-terminated steps.