Refactor `image-processing` unit tests
So far, the unit tests required container registry credentials to run and therefore
were skipped in the GitHub Action based tests.

Rework test cases that do not rely on image pushs or pulls to not be skipped.

Rework test cases that rely on image processing to work against an in-memory temporary registry.