Improve logging in error case for pod retrieval
During investigtion of a failed build run, we ran into the issue that the error
message could be obtained from the TaskRun, but the pod look-up failed for an
unclear reason. Since there is no logging of the error, there is nothing to see.

Add log line in case the pod cannot be looked-up.

Add namespace to another error logging to be more clear in the logs.

Rename `v1` import to `corev1` to be more aligned with other sources.

Sort imports by have Kubernetes imports separately.