Reliably detect evicted pods for failing BuildRuns
When a BuildRun pod is failing it is not necessary to check for failedContainer first.
We can directly check if the pod is evicted and reflect that in the BuildRun condition
Depending on how the system is evicting the pod, `extractFailedPodAndContainer` might return a non-nil value for `failedContainer`.
In such scenarios we were not detecting if the pod was evicted.
This is now fixed as we first check the Reason