Tweak file listing to omit `.git` directory
In almost all cases, the `.git` directory content is not directly helpful
in the file listing and it leads to a lot of noise in the output.

Do not show `.git` files in file listing, but show a summary of size and
file count at the end in the footer section.

Signed-off-by: Matthias Diester <matthias.diester@de.ibm.com>