Add a new filed to describe resource kind for executing buildrun feat(api): Introduce BuildExecutor for multiple runner kinds This commit introduces a new BuildExecutor field to the BuildRunStatus to define the resource responsible for executing a build. This change allows for supporting different kinds of runners, such as TaskRun and PipelineRun.
The new BuildExecutor struct contains two fields:
Name: The name of the runner resource (e.g., the TaskRun or PipelineRun name).
Kind: The kind of the runner resource (e.g., "TaskRun", "PipelineRun").
With this change, the TaskRunName field is now deprecated.

Conversion Logic:
v1beta1 → v1alpha1:
The ImageBuildRun.Name is mapped to the LatestTaskRunRef field in v1alpha1. If ImageBuildRun is not present, it falls back to the deprecated TaskRunName field.
The Kind information is not passed, as v1alpha1 only supports TaskRuns.
v1alpha1 → v1beta1:
The LatestTaskRunRef from v1alpha1 is mapped to ImageBuildRun.Name.
ImageBuildRun.Kind is defaulted to "TaskRun" for compatibility.

Added `// nolint:staticcheck` a directive to disable a specific linter check for a particular line or block of code.
:staticcheck: This specifies the name of the linter to disable. staticcheck is a popular Go linter that performs a wide variety of static analysis checks.

Signed-off-by: Hasan Awad <hasan.m.awad94@gmail.com>